Make functions public, move exhaustive tests to separate suite
