+++ src/apps/benchmarks/benchmarkVertex.c
@@ -39,12 +39,13 @@ BEGIN_BENCHMARKS();
 H3Index* vertexes = calloc(6, sizeof(H3Index));
 
 BENCHMARK(getCellVertexes, 10000, {
-    for (int i = 0; i < ring2Count; i++) getCellVertexes(ring2[i], vertexes);
+    for (int i = 0; i < ring2Count; i++)
+        H3_EXPORT(getCellVertexes)(ring2[i], vertexes);
 });
 
 BENCHMARK(getCellVertexesPent, 10000, {
     for (int i = 0; i < ring2PentCount; i++)
-        getCellVertexes(ring2Pent[i], vertexes);
+        H3_EXPORT(getCellVertexes)(ring2Pent[i], vertexes);
 });
 
 free(vertexes);
