+++ src/h3lib/include/vertex.h
@@ -40,8 +40,5 @@ typedef struct {
 
 int vertexNumForDirection(const H3Index origin, const Direction direction);
 Direction directionForVertexNum(const H3Index origin, const int vertexNum);
-H3Index getCellVertex(H3Index origin, int vertexNum);
-void getCellVertexes(H3Index origin, H3Index* vertexes);
-void vertexToPoint(H3Index vertex, GeoCoord* coord);
 
 #endif
